pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ethos"
version = "0.1.0"
description = "Principle-based safety gates and Signed Integrity Graph CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "pyyaml>=6.0.1",
  "cryptography>=42.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
ethos = "ethos.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
